cleanup creation dom handler
